+++ packages/viz/src/components/settings/CgmTable.js
@@ -24,17 +24,35 @@ class CgmTable extends React.Component {
     const cgmSensorExpirationDate = this.formatDate(cgm.expirationDate);
 
     return (
-      <table className={styles.cgmTable}>
-        <caption className={styles.bdlgSettingsHeader}>
+      <table id="settings-table-cgm" className={styles.cgmTable}>
+        <caption id="settings-table-cgm-title" className={styles.bdlgSettingsHeader}>
           {t('CGM')}
         </caption>
         <tbody>
-          <tr><td>{t('Manufacturer')}</td><td>{cgm.manufacturer}</td></tr>
-          <tr><td>{t('Product')}</td><td>{cgm.name}</td></tr>
-          <tr><td>{t('Cgm sensor expiration date')}</td><td>{cgmSensorExpirationDate}</td></tr>
-          <tr><td>{t('Cgm transmitter software version')}</td><td>{cgm.swVersionTransmitter}</td></tr>
-          <tr><td>{t('Cgm transmitter id')}</td><td>{cgm.transmitterId}</td></tr>
-          <tr><td>{t('Cgm transmitter end of life')}</td><td>{cgmEndOfLife}</td></tr>
+          <tr>
+            <td id="settings-table-cgm-manufacturer">{t('Manufacturer')}</td>
+            <td id="settings-table-cgm-manufacturer-value">{cgm.manufacturer}</td>
+          </tr>
+          <tr>
+            <td id="settings-table-cgm-product">{t('Product')}</td>
+            <td id="settings-table-cgm-product-value">{cgm.name}</td>
+          </tr>
+          <tr>
+            <td id="settings-table-cgm-expdate">{t('Cgm sensor expiration date')}</td>
+            <td id="settings-table-cgm-expdate-value" data-isodate={cgm.expirationDate}>{cgmSensorExpirationDate}</td>
+          </tr>
+          <tr>
+            <td id="settings-table-cgm-swversion">{t('Cgm transmitter software version')}</td>
+            <td id="settings-table-cgm-swversion-value">{cgm.swVersionTransmitter}</td>
+          </tr>
+          <tr>
+            <td id="settings-table-cgm-transmitterid">{t('Cgm transmitter id')}</td>
+            <td id="settings-table-cgm-transmitterid-value">{cgm.transmitterId}</td>
+          </tr>
+          <tr>
+            <td id="settings-table-cgm-eoldate">{t('Cgm transmitter end of life')}</td>
+            <td id="settings-table-cgm-eoldate-value" data-isodate={cgm.endOfLifeTransmitterDate}>{cgmEndOfLife}</td>
+          </tr>
         </tbody>
       </table>
     );