+++ packages/viz/src/utils/settings/diabeloopData.js
@@ -38,6 +38,7 @@ export function getParametersByLevel(parameters) {
 
       const value = formatParameterValue(parameter.value, parameter.unit);
       const param = {
+        rawData: parameter.name,
         name: t(`params|${parameter.name}`),
         value,
         unit: parameter.unit,