+++ packages/yourloops/pages/signup/signup-radio-label.tsx
@@ -30,6 +30,7 @@ import { useTranslation } from "react-i18next";
 import Typography from "@material-ui/core/Typography";
 
 interface RadioLabelProps {
+  id: string;
   header: string;
   body: string;
 }
@@ -39,10 +40,10 @@ export default function RadioLabel(props: RadioLabelProps): JSX.Element {
 
   return (
     <React.Fragment>
-      <Typography variant="h6" gutterBottom>
+      <Typography id={`${props.id}-header`} variant="h6" gutterBottom>
         {t(props.header)}
       </Typography>
-      <Typography variant="body2" gutterBottom>
+      <Typography id={`${props.id}-body`} variant="body2" gutterBottom>
         {t(props.body)}
       </Typography>
     </React.Fragment>