+++ CHANGELOG.md
@@ -23,6 +23,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-816 Add JUnit reports to Jenkins
 - YLP-832 Add TU for zendesk manager
 - YLP-836 Add TU for lib/auth
+- YLP-887 Add more HTML selectors to ease system tests
 
 ## 2.0.2 - 2021-07-05
 ### Fixed